Test: 프론트 E2E smoke 시나리오를 여정별로 분리
